Tested input and output amounts

+++ BitcoinBook/TransactionOutput.cs
@@ -4,10 +4,10 @@ namespace BitcoinBook
 {
     public class TransactionOutput
     {
-        public ulong Amount { get; }
+        public int Amount { get; }
         public ScriptPubKey ScriptPubKey { get; }
 
-        public TransactionOutput(ulong amount, ScriptPubKey scriptPubKey)
+        public TransactionOutput(int amount, ScriptPubKey scriptPubKey)
         {
             Amount = amount;
             ScriptPubKey = scriptPubKey ?? throw new ArgumentNullException(nameof(scriptPubKey));
